# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LeasePro API - RAG-powered lease intelligence platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document and its chunks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Document:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        fileType:
          type: string
        status:
          type: string
          enum: [pending, processing, ready, error]
        chunkCount:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        uploadedAt:
          type: string
          format: date-time
        processedAt:
          type: ["string", "null"]
          format: date-time
        folderId:
          type: ["integer", "null"]
        source:
          type: string
          enum: [upload, sharepoint]
        relativePath:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [all, restricted]
      required:
        - id
        - filename
        - originalName
        - fileType
        - status
        - uploadedAt
...